+++ NEWS.md
@@ -1,5 +1,11 @@
 # sjmisc 2.3.1.9000
 
+## General
+
+* Argument `value` in `set_na()` is deprecated. Please use `na` instead.
+* Argument `recodes` in `rec()` is deprecated. Please use `rec` instead.
+* Argument `lab` in `set_label()` is deprecated. Please use `label` instead.
+
 # sjmisc 2.3.1
 
 ## General